Fixed warnings in the project (implicit cast, move, unused and uninitialized variables).

+++ src/oatpp/web/protocol/http/Http.cpp
@@ -299,7 +299,7 @@ void Parser::parseResponseStartingLine(ResponseStartingLine& line,
     return;
   }
   
-  line.statusCode = caret.parseInt();
+  line.statusCode = (int)caret.parseInt();
 
   auto descriptionLabel = caret.putLabel();
   if(caret.findRN()){
